pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "reconnoiter"
version = "0.0.1"
requires-python= ">=3.10"
dependencies = [
    "arglite",
    "anthropic",
    "pytest",
    "pygithub",
    "gatorgrade",
    "GitPython",
    "requests",
    "python-dotenv",
]
authors = [
    {name = "Douglas Luman", email = "dluman@allegheny.edu"}
]
readme = "README.md"
description = "Tooling to synthesize grading output of gatorgrade for speedy bulk-grading of assignments."

[project.scripts]
recon = "reconnoiter:main.main"
